requirements.txt
flask
opencv-python-headless
torch
torchvision
mediapipe
ultralytics
pillow
gunicorn
numpy
